+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -202,7 +202,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KEV8PWWZ9ASB1DSVHNC1EJRF"),
+  _id: z.string().default("01KEVDRRWGCS1017RNR3C9Y1T1"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -245,7 +245,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KEV8PWWZ9ASB1DSVHNC1EJRF"),
+  id: z.string().default("01KEVDRRWGCS1017RNR3C9Y1T1"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,